satefy guard

+++ rosen/modules/render_service/core/drawable/rs_surface_render_node_drawable.cpp
@@ -1037,8 +1037,9 @@ void RSSurfaceRenderNodeDrawable::DealWithSelfDrawingNodeBuffer(
 
 bool RSSurfaceRenderNodeDrawable::RecordTimestamp(NodeId id, uint32_t seqNum)
 {
-    uint64_t currentTime = std::chrono::duration_cast<std::chrono::nanoseconds>(
-        std::chrono::steady_clock::now().time_since_epoch()).count();
+    uint64_t currentTime = static_cast<uint64_t>(
+        std::chrono::duration_cast<std::chrono::nanoseconds>(
+        std::chrono::steady_clock::now().time_since_epoch()).count());
     auto& surfaceFpsManager = RSSurfaceFpsManager::GetInstance();
     return surfaceFpsManager.RecordPresentTime(id, currentTime, seqNum);
 }
